+++ include/screeps/path-finder.h
@@ -15,26 +15,26 @@ struct path_finder_t {
 	using callback_t = std::function<const cost_matrix_t*(room_location_t)>;
 
 	struct options_t {
-		uint8_t plain_cost = 1;
-		uint8_t swamp_cost = 5;
+		int plain_cost = 1;
+		int swamp_cost = 5;
 		bool flee = false;
-		uint16_t max_ops = 2000;
-		uint8_t max_rooms = 16;
-		uint32_t max_cost = 0x7fffffff;
+		int max_ops = 2000;
+		int max_rooms = 16;
+		int max_cost = 0x7fffffff;
 		double heuristic_weight = 1.2;
 		callback_t room_callback = nullptr;
 	};
 
 	struct result_t {
 		path_ptr_t path;
-		uint32_t ops;
-		uint32_t cost;
+		int32_t ops;
+		int32_t cost;
 		bool incomplete;
 	};
 
 	struct goal_t {
 		position_t pos;
-		uint8_t range;
+		int32_t range;
 	};
 	using goals_t = std::vector<goal_t>;
